t2v-requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
args==0.1.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
braceexpand==0.1.7
cachetools==5.5.1
certifi==2025.4.26
charset-normalizer==3.4.2
clint==0.5.1
# clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.1
cycler==0.12.1
datasets==3.5.1
decorator==4.4.2
diffusers==0.30.0
dill==0.3.8
distro==1.9.0
einops==0.8.1
exceptiongroup==1.2.2
fairscale==0.4.13
filelock==3.17.0
fire==0.4.0
fonttools==4.55.8
frozenlist==1.6.0
fsspec==2025.3.0
ftfy==6.3.1
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hpsv2==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
image-reward==1.5
imageio-ffmpeg==0.4.9
importlib_metadata==8.7.0
iniconfig==2.1.0
iopath==0.1.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
moviepy==1.0.3
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.4.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvitop==1.4.1
omegaconf==2.3.0
open_clip_paddlepaddle-gpu==2.32.0
openai==1.77.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
portalocker==3.1.1
proglog==0.1.10
propcache==0.3.1
protobuf==3.20.3
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
PySocks==1.7.1
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
t2v_metrics==1.2
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==0.6.13
tokenizers==0.15.2
tomli==2.2.1
paddlepaddle-gpu==2.7.0
paddlepaddle-gpuaudio==2.7.0
paddlepaddle-gpuvision==0.22.0
tqdm==4.67.1
transformers==4.36.1
triton==3.3.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.69
ultralytics-thop==2.0.14
urllib3==2.4.0
wcwidth==0.2.13
webdataset==0.2.111
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
